change the array of coroutines into a unordered_map

+++ coroutine.h
@@ -1,6 +1,8 @@
 #ifndef C_COROUTINE_H
 #define C_COROUTINE_H
 
+#include <stdint.h>
+
 #define COROUTINE_DEAD 0
 #define COROUTINE_READY 1
 #define COROUTINE_RUNNING 2
@@ -13,9 +15,9 @@ typedef void (*coroutine_func)(schedule *, void *ud);
 schedule *coroutine_open(void);
 void coroutine_close(schedule *);
 
-int coroutine_new(schedule *, coroutine_func, void *ud);
-void coroutine_resume(schedule *, int id);
-int coroutine_status(schedule *, int id);
+int64_t coroutine_new(schedule *, coroutine_func, void *ud);
+void coroutine_resume(schedule *, int64_t id);
+int coroutine_status(schedule *, int64_t id);
 int coroutine_running(schedule *);
 void coroutine_yield(schedule *);
